+++ src/collective/exportimport/import_content.py
@@ -80,7 +80,7 @@ def get_absolute_blob_path(obj, blob_path):
 
 def filesystem_walker(path=None):
     root = Path(path)
-    assert(root.is_dir())
+    assert root.is_dir()
 
     # first import json-files directly in the path
     json_files = [i for i in root.glob("*.json") if i.stem.isdecimal()]
@@ -104,7 +104,6 @@ def filesystem_walker(path=None):
 
 
 class ImportContent(BrowserView):
-
     template = ViewPageTemplateFile("templates/import_content.pt")
 
     # You can specify a default-target container for all items of a type.
@@ -171,7 +170,7 @@ class ImportContent(BrowserView):
             # This is an error.  But when you upload 10 GB AND select a server file,
             # it is a pity when you would have to upload again.
             api.portal.show_message(
-                _(u"json file was uploaded, so the selected server file was ignored."),
+                _("json file was uploaded, so the selected server file was ignored."),
                 request=self.request,
                 type="warn",
             )
@@ -207,7 +206,7 @@ class ImportContent(BrowserView):
                 status = "error"
                 msg = str(e)
                 api.portal.show_message(
-                    _(u"Exception during upload: {}").format(e),
+                    _("Exception during upload: {}").format(e),
                     request=self.request,
                 )
             else:
@@ -243,7 +242,7 @@ class ImportContent(BrowserView):
 
     def commit_hook(self, added, index):
         """Hook to do something after importing every x items."""
-        msg = u"Committing after creating {} of {} handled items...".format(
+        msg = "Committing after creating {} of {} handled items...".format(
             len(added), index
         )
         logger.info(msg)
@@ -301,10 +300,10 @@ class ImportContent(BrowserView):
         added = self.import_new_content(data)
         end = datetime.now()
         delta = end - start
-        msg = u"Imported {} items".format(len(added))
+        msg = "Imported {} items".format(len(added))
         transaction.get().note(msg)
         transaction.commit()
-        msg = u"{} in {} seconds".format(msg, delta.seconds)
+        msg = "{} in {} seconds".format(msg, delta.seconds)
         logger.info(msg)
         return msg
 
@@ -325,7 +324,9 @@ class ImportContent(BrowserView):
             if not self.should_include(item_path):
                 return False
             elif self.should_drop(item_path):
-                logger.info(u"Skipping %s, even though listed in INCLUDE_PATHS", item_path)
+                logger.info(
+                    "Skipping %s, even though listed in INCLUDE_PATHS", item_path
+                )
                 return False
         else:
             if self.should_drop(item_path):
@@ -337,9 +338,9 @@ class ImportContent(BrowserView):
         added = []
 
         if getattr(data, "len", None):
-            logger.info(u"Importing {} items".format(len(data)))
+            logger.info("Importing {} items".format(len(data)))
         else:
-            logger.info(u"Importing data")
+            logger.info("Importing data")
         for index, item in enumerate(data, start=1):
             if self.limit and len(added) >= self.limit:
                 break
@@ -358,7 +359,7 @@ class ImportContent(BrowserView):
             new_id = unquote(item["@id"]).split("/")[-1]
             if new_id != item["id"]:
                 logger.info(
-                    u"Conflicting ids in url ({}) and id ({}). Using {}".format(
+                    "Conflicting ids in url ({}) and id ({}). Using {}".format(
                         new_id, item["id"], new_id
                     )
                 )
@@ -386,7 +387,7 @@ class ImportContent(BrowserView):
 
             if not container:
                 logger.warning(
-                    u"No container (parent was {}) found for {} {}".format(
+                    "No container (parent was {}) found for {} {}".format(
                         item["parent"]["@type"], item["@type"], item["@id"]
                     )
                 )
@@ -394,7 +395,7 @@ class ImportContent(BrowserView):
 
             if not getattr(aq_base(container), "isPrincipiaFolderish", False):
                 logger.warning(
-                    u"Container {} for {} is not folderish".format(
+                    "Container {} for {} is not folderish".format(
                         container.absolute_url(), item["@id"]
                     )
                 )
@@ -408,7 +409,7 @@ class ImportContent(BrowserView):
                 if self.handle_existing_content == 0:
                     # Skip
                     logger.info(
-                        u"{} ({}) already exists. Skipping it.".format(
+                        "{} ({}) already exists. Skipping it.".format(
                             item["id"], item["@id"]
                         )
                     )
@@ -417,7 +418,7 @@ class ImportContent(BrowserView):
                 elif self.handle_existing_content == 1:
                     # Replace content before creating it new
                     logger.info(
-                        u"{} ({}) already exists. Replacing it.".format(
+                        "{} ({}) already exists. Replacing it.".format(
                             item["id"], item["@id"]
                         )
                     )
@@ -426,7 +427,7 @@ class ImportContent(BrowserView):
                 elif self.handle_existing_content == 2:
                     # Update existing item
                     logger.info(
-                        u"{} ({}) already exists. Updating it.".format(
+                        "{} ({}) already exists. Updating it.".format(
                             item["id"], item["@id"]
                         )
                     )
@@ -438,7 +439,7 @@ class ImportContent(BrowserView):
                     duplicate = item["id"]
                     item["id"] = "{}-{}".format(item["id"], random.randint(1000, 9999))
                     logger.info(
-                        u"{} ({}) already exists. Created as {}".format(
+                        "{} ({}) already exists. Created as {}".format(
                             duplicate, item["@id"], item["id"]
                         )
                     )
@@ -466,16 +467,17 @@ class ImportContent(BrowserView):
                 if self.commit and not len(added) % self.commit:
                     self.commit_hook(added, index)
             except Exception as e:
-                item_id = item['@id'].split('/')[-1]
+                item_id = item["@id"].split("/")[-1]
                 container.manage_delObjects(item_id)
                 logger.warning(e)
-                logger.warning("Didn't add %s %s", item["@type"], item["@id"], exc_info=True)
+                logger.warning(
+                    "Didn't add %s %s", item["@type"], item["@id"], exc_info=True
+                )
                 continue
 
         return added
 
     def handle_new_object(self, item, index, new):
-
         new, item = self.global_obj_hook_before_deserializing(new, item)
 
         # import using plone.restapi deserializers
@@ -484,13 +486,15 @@ class ImportContent(BrowserView):
             try:
                 new = deserializer(validate_all=False, data=item)
             except TypeError as error:
-                if 'unexpected keyword argument' in str(error):
+                if "unexpected keyword argument" in str(error):
                     self.request["BODY"] = json.dumps(item)
                     new = deserializer(validate_all=False)
                 else:
                     raise error
         except Exception:
-            logger.warning("Cannot deserialize %s %s", item["@type"], item["@id"], exc_info=True)
+            logger.warning(
+                "Cannot deserialize %s %s", item["@type"], item["@id"], exc_info=True
+            )
             raise
 
         # Blobs can be exported as only a path in the blob storage.
@@ -506,9 +510,7 @@ class ImportContent(BrowserView):
             # Happens only when we import content that doesn't have a UID
             # for instance when importing from non Plone systems.
             logger.info(
-                "Created new UID for item %s with type %s.",
-                item["@id"],
-                item["@type"]
+                "Created new UID for item %s with type %s.", item["@id"], item["@type"]
             )
             item["UID"] = uuid
 
@@ -534,9 +536,7 @@ class ImportContent(BrowserView):
             new.creation_date = creation_date
             new.aq_base.creation_date_migrated = creation_date
         logger.info(
-            "Created item #{}: {} {}".format(
-                index, item["@type"], new.absolute_url()
-            )
+            "Created item #{}: {} {}".format(index, item["@type"], new.absolute_url())
         )
         return new
 
@@ -598,7 +598,12 @@ class ImportContent(BrowserView):
             try:
                 new = deserializer(validate_all=False, data=version)
             except Exception:
-                logger.warning("Cannot deserialize %s %s", item["@type"], item["@id"], exc_info=True)
+                logger.warning(
+                    "Cannot deserialize %s %s",
+                    item["@type"],
+                    item["@id"],
+                    exc_info=True,
+                )
                 return
 
             self.save_revision(new, version, initial)
@@ -609,7 +614,9 @@ class ImportContent(BrowserView):
         try:
             new = deserializer(validate_all=False, data=item)
         except Exception:
-            logger.warning("Cannot deserialize %s %s", item["@type"], item["@id"], exc_info=True)
+            logger.warning(
+                "Cannot deserialize %s %s", item["@type"], item["@id"], exc_info=True
+            )
             return
 
         self.import_blob_paths(new, item)
@@ -671,7 +678,7 @@ class ImportContent(BrowserView):
         from plone.app.versioningbehavior import _ as PAV
 
         if initial:
-            comment = PAV(u"initial_version_changeNote", default=u"Initial version")
+            comment = PAV("initial_version_changeNote", default="Initial version")
         else:
             comment = item.get("changeNote")
         sys_metadata = {
@@ -778,15 +785,12 @@ class ImportContent(BrowserView):
             return
         constrains.setConstrainTypesMode(ENABLED)
 
-        locally_allowed_types = item["exportimport.constrains"][
-            "locally_allowed_types"
-        ]
+        locally_allowed_types = item["exportimport.constrains"]["locally_allowed_types"]
         try:
             constrains.setLocallyAllowedTypes(locally_allowed_types)
         except ValueError:
             logger.warning(
-                "Cannot setLocallyAllowedTypes on %s", item["@id"],
-                exc_info=True
+                "Cannot setLocallyAllowedTypes on %s", item["@id"], exc_info=True
             )
 
         immediately_addable_types = item["exportimport.constrains"][
@@ -796,8 +800,7 @@ class ImportContent(BrowserView):
             constrains.setImmediatelyAddableTypes(immediately_addable_types)
         except ValueError:
             logger.warning(
-                "Cannot setImmediatelyAddableTypes on %s", item["@id"],
-                exc_info=True
+                "Cannot setImmediatelyAddableTypes on %s", item["@id"], exc_info=True
             )
 
     def import_review_state(self, obj, item):
@@ -884,7 +887,7 @@ class ImportContent(BrowserView):
             container = api.content.get(path=container_path)
             if not container:
                 raise RuntimeError(
-                    u"Target folder {} for type {} is missing".format(
+                    "Target folder {} for type {} is missing".format(
                         container_path, item["@type"]
                     )
                 )
@@ -945,11 +948,20 @@ class ImportContent(BrowserView):
             return api.portal.get().__parent__
 
         if item["parent"]["@type"] == "Plone Site":
-            return api.portal.get()
+            # if view is called on a folder, use current context ad root
+            # if view is called on Site Root, use current context too
+            return self.context
 
         # If the item is missing look for a item with the path of the old parent
         parent_url = unquote(item["parent"]["@id"])
         parent_path = urlparse(parent_url).path
+
+        if self.context.portal_type != "Plone Site":
+            # fix parent_path to search created contents into current context
+            context_path = "/".join(self.context.getPhysicalPath())
+            parent_rootless = "/".join(parent_path.split("/")[2:])
+            parent_path = f"{context_path}/{parent_rootless}"
+
         # physical path is bytes in Zope 2 (not in Zope 4)
         # so we need to encode parent_path before using plone.api.content.get
         if isinstance(self.context.getPhysicalPath()[0], bytes):
@@ -996,7 +1008,9 @@ class ImportContent(BrowserView):
 
         # Handle folderish Documents provided by plone.volto
         fti = getUtility(IDexterityFTI, name="Document")
-        parent_type = "Document" if fti.klass.endswith("FolderishDocument") else "Folder"
+        parent_type = (
+            "Document" if fti.klass.endswith("FolderishDocument") else "Folder"
+        )
         # create original structure for imported content
         for element in parent_path:
             if element not in folder:
@@ -1006,7 +1020,11 @@ class ImportContent(BrowserView):
                     id=element,
                     title=element,
                 )
-                logger.info(u"Created container %s to hold %s", folder.absolute_url(), item["@id"])
+                logger.info(
+                    "Created container %s to hold %s",
+                    folder.absolute_url(),
+                    item["@id"],
+                )
             else:
                 folder = folder[element]
 
@@ -1041,16 +1059,18 @@ def fix_portal_type(portal_type):
 
 class ResetModifiedAndCreatedDate(BrowserView):
     def __call__(self):
-        self.title = _(u"Reset creation and modification date")
-        self.help_text = _("<p>Creation- and modification-dates are changed during import." \
-                         "This resets them to the original dates of the imported content.</p>")
+        self.title = _("Reset creation and modification date")
+        self.help_text = _(
+            "<p>Creation- and modification-dates are changed during import."
+            "This resets them to the original dates of the imported content.</p>"
+        )
         if not self.request.form.get("form.submitted", False):
             return self.index()
 
         portal = api.portal.get()
 
         portal.ZopeFindAndApply(portal, search_sub=True, apply_func=reset_dates)
-        msg = _(u"Finished resetting creation and modification dates.")
+        msg = _("Finished resetting creation and modification dates.")
         logger.info(msg)
         api.portal.show_message(msg, self.request)
         return self.index()
@@ -1072,12 +1092,16 @@ def reset_dates(obj, path):
 
 class FixCollectionQueries(BrowserView):
     def __call__(self):
-        self.title = _(u"Fix collection queries")
-        self.help_text = _(u"""<p>This fixes invalid collection-criteria that were imported from Plone 4 or 5.</p>""")
+        self.title = _("Fix collection queries")
+        self.help_text = _(
+            """<p>This fixes invalid collection-criteria that were imported from Plone 4 or 5.</p>"""
+        )
 
         if not HAS_COLLECTION_FIX:
             api.portal.show_message(
-                _(u"plone.app.querystring.upgrades.fix_select_all_existing_collections is not available"),
+                _(
+                    "plone.app.querystring.upgrades.fix_select_all_existing_collections is not available"
+                ),
                 self.request,
             )
             return self.index()